backend/requirements.txt
fastapi
uvicorn
groq
huggingface_hub
python-dotenv
motor
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
pymongo
